Initial PoC for an external stager
The `stager` docker image takes a standard Diego staging task
definition, downloads the app, lifecycle builder and buildpacks.
It then runs the builder component of the lifecycle and uploads
the generated droplet back to the cc-uploader.

The files in this commit rely on a further small change to the
stager, to dump the task definition into the log, and also have
some dependencies on HCF (log file location and networking layer).